Updated Multiplication Visualizer to work with Expression and it now works with any form of expression (i.e. there are no restrictions).
For now I'm omitting the labels that give more info about the multiplication i.e. the 2*9 labels on the left and top of the rectangle.

Sizing isn't fully optimized yet either. Would like some suggestions as to how to fix the sizing.

Created a demo class to test the MultiplicationVisualizer.